requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-Limiter>=3.5.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
PyJWT>=2.8.0
bcrypt>=4.0.0
python-magic>=0.4.27
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pytest>=7.4.0
requests>=2.31.0
redis>=4.6.0
celery>=5.3.0
flask-wtf>=1.1.0
openai>=1.25.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11